requirements.txt
asgiref==3.11.0
Django==6.0
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-filter==25.2
requests>=2.31.0